+++ sdk/face/azure-ai-vision-face/src/main/java/com/azure/ai/vision/face/models/FaceAttributeType.java
@@ -69,8 +69,7 @@ public final class FaceAttributeType extends ExpandableStringEnum<FaceAttributeT
      * The overall image quality regarding whether the image being used in the detection is of sufficient quality to
      * attempt face recognition on. The value is an informal rating of low, medium, or high. Only 'high' quality images
      * are recommended for person enrollment and quality at or above 'medium' is recommended for identification
-     * scenarios. The attribute is only available when using any combinations of detection models detection_01 or
-     * detection_03, and recognition models recognition_03 or recognition_04.
+     * scenarios. The attribute is only available when using recognition models recognition_03 or recognition_04.
      */
     @Generated
     public static final FaceAttributeType QUALITY_FOR_RECOGNITION = fromString("qualityForRecognition");
@@ -134,7 +133,7 @@ public final class FaceAttributeType extends ExpandableStringEnum<FaceAttributeT
     /**
      * Face attribute supported by Detection01.
      */
-    public static final class Detection01 {
+    public static final class ModelDetection01 {
 
         /**
          * 3-D roll/yaw/pitch angles for face direction.
@@ -179,26 +178,30 @@ public final class FaceAttributeType extends ExpandableStringEnum<FaceAttributeT
     /**
      * Face attribute supported by Detection03.
      */
-    public static final class Detection03 {
+    public static final class ModelDetection03 {
+
+        /**
+         * Face is blurry or not. Level returns 'Low', 'Medium' or 'High'. Value returns a number between [0,1], the
+         * larger the blurrier.
+         */
+        public static final FaceAttributeType BLUR = FaceAttributeType.BLUR;
 
         /**
          * 3-D roll/yaw/pitch angles for face direction.
          */
-        @Generated
         public static final FaceAttributeType HEAD_POSE = FaceAttributeType.HEAD_POSE;
 
         /**
          * Whether each face is wearing a mask. Mask type returns 'noMask', 'faceMask', 'otherMaskOrOcclusion', or
          * 'uncertain'. Value returns a boolean 'noseAndMouthCovered' indicating whether nose and mouth are covered.
          */
-        @Generated
         public static final FaceAttributeType MASK = FaceAttributeType.MASK;
     }
 
     /**
      * Face attribute supported by Recognition03.
      */
-    public static final class Recognition03 {
+    public static final class ModelRecognition03 {
 
         /**
          * The overall image quality regarding whether the image being used in the detection is of sufficient quality to
@@ -208,14 +211,13 @@ public final class FaceAttributeType extends ExpandableStringEnum<FaceAttributeT
          * scenarios. The attribute is only available when using any combinations of detection models detection_01 or
          * detection_03, and recognition models recognition_03 or recognition_04.
          */
-        @Generated
         public static final FaceAttributeType QUALITY_FOR_RECOGNITION = FaceAttributeType.QUALITY_FOR_RECOGNITION;
     }
 
     /**
      * Face attribute supported by Recognition04.
      */
-    public static final class Recognition04 {
+    public static final class ModelRecognition04 {
 
         /**
          * The overall image quality regarding whether the image being used in the detection is of sufficient quality to
@@ -225,7 +227,6 @@ public final class FaceAttributeType extends ExpandableStringEnum<FaceAttributeT
          * scenarios. The attribute is only available when using any combinations of detection models detection_01 or
          * detection_03, and recognition models recognition_03 or recognition_04.
          */
-        @Generated
         public static final FaceAttributeType QUALITY_FOR_RECOGNITION = FaceAttributeType.QUALITY_FOR_RECOGNITION;
     }
 }