+++ sdk/cosmos/azure-cosmos-spark_3-3_2-12/CHANGELOG.md
@@ -1,10 +1,10 @@
 ## Release History
 
-### 4.32.0-beta.1 (Unreleased)
+### 4.32.0 (2024-05-24)
 
 #### Features Added
-
-#### Breaking Changes
+* Added config option `spark.cosmos.auth.aad.clientCertPemBase64` to allow using SPN (ServicePrincipal name) authentication with certificate instead of client secret. - See [PR 40325](https://github.com/Azure/azure-sdk-for-java/pull/40325)
+* Added config option `spark.cosmos.accountDataResolverServiceName` to allow specifying which `AccountDataResolver` trait implementation to use if there are multiple on the class path. - See [PR 40325](https://github.com/Azure/azure-sdk-for-java/pull/40325)
 
 #### Bugs Fixed
 * Fixed an issue where `SHOW DATABASES IN` only return one database even though multiple databases exist. - See [PR 40277](https://github.com/Azure/azure-sdk-for-java/pull/40277)